Move toElementRows() to RowsTranslationUtil.h
Summary: `toElementRows()` function is originally defined in LambdaFunctionUtil.h/cpp but has been used in several places outside of lambda functions. This diff moves `toElementRows()` to RowsTranslationUtil.h and extracts LambdaFunctionUtil and RowsTranslationUtil into a new target velox_functions_util. This change is needed for a subsequent fix of cast-in-try.

Differential Revision: D39983983

fbshipit-source-id: 14dedc38dd7d03e76f3e150c378ed6bf612928f0