Fix lambda-in-try for lambda functions on non-array and non-map types (#2707)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2707

Lambda expression inside Try expression used to not subpress
errors but throw them immediately because of the inproper setting
of lambdaCtx::throwOnError_. This diff fixes lambda-in-try for
lambda functions that process non-array and non-map types. The
fix for lambda functions that process array or map will come in a
subsequent diff.

This is the first diff to fix https://github.com/facebookincubator/velox/issues/2605.

Differential Revision: https://internalfb.com/D39992324

fbshipit-source-id: acd00d1f6a801d514ec51b8e8cfc650ed4f3ffce