Fix SignatureBinder to handle unmatched number of child types (#2799)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2799

When binding a function signature with given argument types, if a parameter type has child
types, the corresponding argument type should have the same number of children. Before
this fix, SignatureBinder::tryBind() throw an exception in this situation. It should instead
return false because trying to bind with unmatched argument types is an expected use case.

Differential Revision: https://internalfb.com/D40203036

fbshipit-source-id: d75571ee135376d8f5ddb0a9fb7acecec7f1b338