Fix SignatureBinder to handle unmatched number of child types (#2799)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2799

When binding a function signature with given argument types, if a parameter type has child
types, the corresponding argument type should have the same number of children. Before
this fix, SignatureBinder::tryBind() throw an exception in this situation. It should instead
return false because trying to bind with unmatched argument types is an expected use case.

Reviewed By: xiaoxmeng

Differential Revision: D40203036

fbshipit-source-id: 717b21ba0d21fef7102f810f375cb98243648add