[Velox] Pass write protocol to table writer
Refactor TableWriter to take WriteProtocol. WriteProtocol can be
used to to configure write and commit, which will be applied
in the change that follows.

Pass WriteProtocol to DataSink that is used by TableWriter.
DataSink can get the WriterParameters for writers from WriteProtocol,
keep track of them  and includes them in the ConnectorCommitInfo.
TableWriter assembles CommitInfo with ConnectorCommitInfo and more
generic info for commit().

TableWriter getOutputs() also calls on WriteProtocol commit() to perform
protocol particular commit() actions. It would also include
the particular format of outputs that commit() returns.

Make the Velox example ScanAndSort.cpp use HiveNoCommitWriteProtocol,
with no behavior change.