Fix and make Expr::addNulls() a static function for code reuse
Summary:
Expr::addNulls() provides a common logic for adding nulls to a vector and handles
all kinds of cases of input. It is useful when we evaluate an expression on only non-
null rows and want to add remaining nulls to the result. This logic is needed in
CastExpr too, so we make it a static function to make the code reuse possible.

Differential Revision: D40530240

fbshipit-source-id: 2175afc377a2acd7b80f9f2596c15a3c68e44fc1