Fix cast-in-try caused by mismatch between element rows and top-level rows (#2682)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2682

The evaluation of cast of array or map vectors process their element vectors
directly. If there are errors during processing, exceptions are kept in
EvalCtx::errors_ at the same rows as the elements that cause errors. If the
cast is wrapped in a try expression, these exceptions are then turned into
nulls at the same rows in the result vector. This lead to a mismatch because
exceptions are kept at element rows whereas values in the result vector are
at top-level rows. This diff transform EvalCtx::errors_ to map exceptions back
to top-level rows at the end of processing array or map vectors.

Reviewed By: mbasmanova

Differential Revision: D39901232

fbshipit-source-id: 53c26b1cd15fc32ea2732a03bf8c3516ab413c3c