Move toElementRows() to RowsTranslationUtil.h (#2703)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2703

`toElementRows()` function is originally defined in LambdaFunctionUtil.h/cpp
but has been used in several places outside of lambda functions. This diff moves
`toElementRows()` to RowsTranslationUtil.h and extracts LambdaFunctionUtil
and RowsTranslationUtil into a new target velox_functions_util. This change is
needed for a subsequent fix of cast-in-try.

Differential Revision: https://internalfb.com/D39983983

fbshipit-source-id: d99254956bac60803aa751c8c783d085fd765e52