Generate not null and unique map keys on VectorFuzzer
Summary:
VectorFuzzer now provides a normalizeMapKeys flags. If true, generated
maps are ensured to have unique and non-null keys. This works by either
generating not null vectors (or checking if the provided key vectors are
not-nullable).
.
To guarantee keys are unique, each map element is checked and cut short if
duplicate values are found.

Reviewed By: mbasmanova

Differential Revision:
D40787363

LaMa Project: L1134559

fbshipit-source-id: e8b62808e6c4e9724da581a5686630d0a1a47b2a