Add a unit test and fix for the task not found issue (#3009)
This is the case where the downstream prematurely aborts
and calls abortTask of the upstream. The abortTask
delete the task asynchronously from the buffers. This causes
and exception due to VELOX_CHECK in the following call tree
(because the buffers does not have the entry for the taskId any more)

Driver.runInternal
  -> PartitionOutput.getOutput()
  -> Destination.flush() (if (blockedDestination) {} path)
  -> PartitionedOutputBufferManager.enqueue()
  -> PartitionedOutputBufferManager.getBuffer()  <- Exception here

In the case I've seen, the abort would happen between the flush() and getBuffer()

I also added a simple fix (tested in our cluster) which removes the VELOX_CHECK
and has the clients check for nullptr.