Mitigate POISONED by IO_THROTTLED: WSE_SERVICE_OUT_OF_RESOURCE
Summary:
For some reason, being investigated in T136142244, our max stripe size ends up being up to 3x the value set in this property, causing service out of resource.

Let's set it to 256MB/3=85MB -> ~72MB (an ideal size for writing to WS) while we fix this to mitigate the problem for everybody.

Differential Revision: D40857332

fbshipit-source-id: d84c912d9305fac61f800093d7aaa5542b3b1f50