Add a fix for the task not found issue (#3009)
This is the case where the downstream prematurely aborts
and calls abortTask of the upstream. The abortTask
delete the task asynchronously from the buffers. This causes
and exception due to VELOX_CHECK in the following call tree
(because the buffers does not have the entry for the taskId any more)

Driver.runInternal
  -> PartitionOutput.getOutput()
  -> Destination.flush() (if (blockedDestination) {} path)
  -> PartitionedOutputBufferManager.enqueue()
  -> PartitionedOutputBufferManager.getBuffer()  <- Exception here

In the case I've seen, the abort would happen between the flush() and getBuffer()

The fix is to delete the code that deletes the buffer for the task. The buffer for
the task will be removed in Task::terminate() as part of the normal Task lifecycle.