Data race in `variant` (#2966)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2966

# Problem
There is a data race in using `toJson()` method in `variant`. It accesses `getOpts()` function to access the singleton and __update the value__. If this happens in multiple threads, this update can happen simultaneously.

We are updating to a fixed value, so I don't think it would break anything. But this is unnecessary risk, and I'm not sure why it even returns modifiable reference.

Let's initialize it with the construction of static memory, and prevent any modification.

Reviewed By: kKPulla

Differential Revision: D40725553

fbshipit-source-id: 2525ad276711e2af27b1241be2a621f07a2a3ac7