Fix all nulls in in list
An in list in Presto may consist of a single null. If this is the case
the predicate is always null for any input. A in list with nulls only
is OK, a values list with no elements is an error.

Adds a signature with unknown type for the case of an all null in
list.