Backtrack when arg type is an invalid map with unknown keys.
Summary:
solve issues  #3046
The fuzzer works as the following:
1. pick a return type.
2. find an expression that can return this type.
3. determine types of args.
4. repeat for type of args.

Consider a situation where 1 selected Array<unknown>, in that
case 2 could be  map_keys function since it returns T and T can
bound to Array<unknown>.

Then 3 would determine that the input should be map<unknown, X>
and hence that would be an invalid input.

This diff capture such thing and back track so a different expression
can be selected.

We can potentially move the checks to chooseRandomSignatureTemplate as well,
but i found this more cleaner.
The other approach will look like this: D40923723

Differential Revision: D40923057

fbshipit-source-id: a86c14fd3f512b8a9d40bcd6b05e4c5a69327e78