Fix map_keys and map_values presto function
Currently both these functions generate a result array of size equal
to the total size of the selectivity vector passed to it. However if
the last valid row in the selectivity vector is smaller than its
total size (that is rows.end() < rows.size()) then this can be
larger than the size of the underlying buffers and can result in an
inconsistent state where the size of the offset or sizes buffer is
insufficient to hold all indices. This patch fixes the
aforementioned bug.

Test Plan
Added a unit test that would fail without this fix
(but needs #3078 for consistency check)