Refactor ParquetData
To prepare for the ARRAY reader, this commit includes the following:

1. Rename FormatData to FormatDataReader, DwrfData to DwrfDataReader,
ParquetData to ParquetDataReader. This is to differenciate from future
nested data structures.

2. Extract a ParquetDataReader super class which handles RowGroup
related operations. The ParquetPrimitiveDataReader become a subclass
of it. This is to leave space for the upcoming
ParquetNestedPrimitiveDataReader.

3. Add ParquetPage structs and added new PageReader::readNextPage()
function that returns a single next ParquetPage. Rename PageReader::
readNextPage(int64_t row) to PageReader::seekToPage(int64_t row) to
differenticiate from it.