Handle Vector function setting errors without allocating result (#3086)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3086

Some Vector functions, e.g. LikeWithRe2, can throw
exceptions for every row and then not set the result VectorPtr.

There are many points within the code where we assume that
result is not a nullptr, e.g. the lines immediately following
vectorFunction_->apply in Expr::applyFunction.  This can lead
to crashes in Velox.

To handle this, I added a check that if the function did not set
result, we set it to be a NULL constant Vector.

For completeness, I also checked if the function did not set
errors for all rows in this case, if  it did not, we set an error for
those rows remaining indicating the function is buggy.

Since at this point all rows should have an error associated
with them it's safe to set the result to be a NULL constant.

This was detected with the ExpressionFuzzer.

Differential Revision: D40992848

fbshipit-source-id: 78446c198186add413cbf4f22bab6d68cd216dde