Move the ownership of driver executor out of QueryCtx (#3111)
Summary:
There is a deadlock in async Task destruction code path:
The last driver destruction runs executed by the driver thread
triggers the task destruction. The task destruction drops the
last reference on the query context. The query context destruction
drops the last reference on the driver thread pool. The driver
thread pool destruction blocks to wait for the queued callbacks to
finish and the driver destruction is current running callback. So
this forms a circle dependency and task destruction hangs forever.
This PR fixes the problem by moving the ownership of the driver
thread pool out of query context and let the user like Presto CPP
to own the driver executor so the task destruction will not block on
the driver executor's destruction to break the dependency circle.

- Remove QueryCtx::createForTest method
- Remove the hack in Task destructor to reset plan fragment
- Change the warning in testingWaitForAllTasksToBeDeleted to a
   CHECK failure
- Check all tasks to be deleted in each test case derived from
  VectorTestBase
- Some code cleanup CMakefile fix

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3111

Reviewed By: mbasmanova

Differential Revision: D41073275

Pulled By: xiaoxmeng

fbshipit-source-id: ba092c5390a9d7bf0b4a7d650632e3d509384292