Enhance lazy vector generation in vector fuzzer
Currently vector fuzzer exposes an API that allows wrapping an input
vector with a lazy layer. If an input vector with multiple
dictionary layers is wrapped in this lazy layer then when it is
loaded, it combines all the dictionary layers into one similar to
what a decoded vector does. This takes away the ExpressionFuzzer's
ability to have lazy inputs with dictionary layers on top.
This change allows the fuzzer to identify this kind of input vector
and only apply the lazy layer on the innermost dictionary layer.

Test Plan:
Modified an existing test case