Add check to enforce invariant related to lazy vectors
This adds an explicit check to ensure an unloaded lazy vectors cannot
be wrapped by two different top level vectors, that is
Dictionary1(lazy1) and Dictionary2(lazy1) is not possible but
Dictionary3(Dictionary1(lazy1)) is possible. This is an attempt
to reduce the possibility of a lazy being loaded with potentially
different set of rows.

Limitation:
This does not prevent the case where another reference to the lazy can
potentially call load on it.

Test Plan:
Added a unit test.