Ensure slice() does not generate overlapping ranges
Summary:
In cases where the first parameter is a constant, but not the second
or third, slice() was generating overlapping (offset, size) ranges, which ends
up breaking other functions upstream that rely on the fact that array elements
do not overlap (like reverse). To prevent these cases, flatten the constant input.

Differential Revision: D41330251

fbshipit-source-id: 6a564fcdab88dadb465125edba12728480402956