Fix incorrect result of string concat() when used in IF expression
Summary:
There is a bug in the string concat() function that leads to incorrect result when
evaluation `if(c0, concat(...), c1)` or `if(c0, f(c1), concat(...))`.
This bug is a misuse of `FlatVector<StringView>::getBufferWithSpace()`
and causes the else-branch to overwrite the results from the then-branch in the if
expression.

When an expression `if(c0, f(c1), g(c2))` is evaluated, `f(c1)` is first
evaluated on all rows where `c0` is true and produces a partial result vector, then
`g(c2)` is evaluated on all the rest of rows and writes to the same result vector.
Therefore, the else-branch must be careful to not overwrite the existing content in
the result vector.

`ConcatFunction::apply()` calls `flatResult->getBufferWithSpace(totalResultBytes)`
to reserve enough space for the concatenation results. However, it doesn't skip
the existing content when writing to it, and doesn't update the buffer size after the
writing. These cause two issues:

When `if(c0, concat(...), c1)` is evaluated and c1's memory pool is
different from the pool of the EvalCtx, the else-branch copies content in c1 to the
partial result vector produced by concat(). But since concat() didn't update the
buffer size, the else-branch thinks the buffer is empty and hence overwrites the
content produced by the concat().

When `if(c0, f(c1), concat(...))` is evaluated, concat() get a buffer via
`getBufferWithSpace()`. `getBufferWithSpace()` may return an existing
buffer of the result vector if it is large enough. However, concat() doesn't skip the
existing content when writing to the buffer, hence it overwrites the results produced
by the then-branch `f(c1)`.

This diff fixes these two issues together and add a unit test that tests both.

Reviewed By: mbasmanova

Differential Revision: D41330939

fbshipit-source-id: 9a52eaf27e3078408cdf8f4ceac9e62a9b3a0454