Prevent approx_most_frequent from crashing due to large buckets (#3259)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3259

In old code we allocate a temporary vector with size equal to the
`buckets` parameter passed by user.  This value can be too large and crash the
process.  As a fix we remove the temporary vector and write directly to the
result buffers.  When allocating the buffers, the memory is capped and will
throw proper exception when this cap is exceeded.

Fix https://github.com/facebookincubator/velox/issues/3101

Reviewed By: mbasmanova

Differential Revision: D41311599

fbshipit-source-id: 00cf67f147cae14dde730245c3e3822a9028df21