Enhance fuzzer to re-use sub-expressions and columns
After this change the expression fuzzer can generated sub-expressions
that re-use some already generated columns. Before picking an input
column to an expression it will check if there are existing columns
with the same type and randomly pick whether to re-use one of them
or create a new one.

Similar functionality is employed to re-use sub-expressions.
However, there is a current limitation to this that only expressions
with no subexpressions are shared.

Test Plan:
Ran fuzzer tests locally and verified manually that columns and
sub-expressions are being re-used.