Fix incorrect results when grouping set keys overlap with aggregate inputs
Summary:
`GroupId` operator currently layouts all the grouping keys, then followed by all aggregate inputs, then the group ID column.  It then sets NULL to some of the key columns according to the currently active grouping set.  There is an edge case that if some columns show up in both keys and inputs, when the downstream aggregate operator receives the data and tries to retrieve the input, it will search by name and find the key column, which is potentially set to NULL.

To avoid this bug, we change the logic when looking up input column to search from backward.  This relies on the fact that `GroupId` always order the key columns before input columns.  This way we can avoid retrieving the column which is set to NULL by `GroupId`.

Fix https://github.com/facebookincubator/velox/issues/3357

Differential Revision: D41533704

fbshipit-source-id: 5dd2619f3047befe42e238e3ba12e66b09a5a36d