Fix template for If expression used in the fuzzer
Summary:
Currently the template for If expression contains one extra input
parameter for each case, the first case (using 3 but expected 2)
works fine because it assumes the extra param is an else clause,
but for the second case (using 4 but expected 3), it expects a
boolean as return type for the third param but gets 'T' and throws
a USER error which both the common and simplified path throw.
Hence it gets ignored as a user error and the fuzzer never crashes.
This patch fixes this bug.

Differential Revision: D41600721

fbshipit-source-id: 99c005e8baee8438cf0f0d1d84eda0ace3d162cd