Add AlphaReaderFactory to velox (#3339)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3339

This change will allow user to register Alpha file format and read them in Velox connectors.  Only non-selective reader is supported for now, and cached input is not supported yet.  The `InputStream` passed to the factory must be a `ReadFileInputStream` with a shared ownership of `ReadFile`.  Later we will refactor the factory to take `ReadFile` directly.

Reviewed By: zzhao0

Differential Revision: D41473280

fbshipit-source-id: 5366b9558ddd1a0878b5dd0c08c152813510e55d