Use ReadFile in InputStreamFactory::create (#3435)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3435

In subsequent changes, `BufferedInput` will only takes `ReadFile` or `ReadFileInputStream`.

As we now use `WSReadFile` in `InputStreamFactory`, all downstream libraries will crash if they do not register an implementation of `BaseStatsReporter`.  The implementation is registered as a `folly::Singleton` and there is currently no way to provide a fallback implementation.  To workaround this, we add a flag `BaseStatsReporter::registered`, and only report the stats from `WSReadFile` when this flag is set to true.  A grep in the codebase suggest only `presto_cpp` will need to set this flag.

Differential Revision: D41749995

fbshipit-source-id: dc56bb677a0975310d8c1cdee1b601e98fd9e5ac