Fix incorrect cast result for constant vector in dictionary encoding (#3424)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3424

Expression fuzzer found a bug in `cast(c0 as SMALLINT)` when `c0` is a constant vector inside a dictionary encoding (https://github.com/facebookincubator/velox/issues/3300). In this situation, CastExpr tries to peel off the encoding before processing the data and wrap the encoding back to the result afterwards. But when we extract the encoding with DecodedVector::dictionaryWrapping(), dictionaryWrapping() only gets the top-level dictionary layer because `VectorEncoding::Simple::CONSTANT` is not considered a wrapper. So this constant layer is peeled off first, but not wrapped back to the result, hence causing the error. This diff fixes this bug by making DecodedVector consider `VectorEncoding::Simple::CONSTANT` as a wrapper and DecodedVector::dictionaryWrapping() to extract the constant encoding.

Reviewed By: Yuhta

Differential Revision: D41450072

fbshipit-source-id: 9f51fc89d1e1682b8cf444860f0de1cb092b509f