Add Parquet struct reader
Each struct or list must have one leaf descendent that reads all the repdefs of its column chunk. Nested structs/lists may share te same leaf. The single leaf for which the whole column chunk of repdefs will be read is selected to be or the narrowest type. Filtered leaves  are preferred over non-filtered.

The mentioned leaves are read for the column chunk on first use when first reading the containing top level list/struct. Other leaf columns are read as needed.

Non-top level leaves for which the repdefs are not read for the column
chunk have the definition levels translated into null flags and leaf
value counts at the time of reading the page header.

The lengths and nulls for all non-leaf readers are set by
ensureRepDefs() when calling read() for the topmost struct/list.  Each
reader will access these when its read() is called.

Nested structs with filters in inner structs may end a read() with
different substructs at different rows. Before initiating a new top
level read, the nested structs and their children are all brought to
the end of the rows covered by the repdefs of the previous read(). In
this way, the new repdefs all pick up at the same top level row.