Add fuzzer support for Switch expression (#3373)
Summary:
Needs to be enabled by adding "switch" to the --special_forms flag.
Currently, it randomly picks between 1 to 5 cases (condition, then
pairs) and also randomly chooses whether to include the final else
clause or not. It also reduces the nesting level when being used to
avoid blowing up the expression tree as it can add up to 11 more
branches to it.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3373

Test Plan: Manually tested and verified.

Reviewed By: kagamiori

Differential Revision: D41597545

Pulled By: bikramSingh91

fbshipit-source-id: cd50365afe1d68c51d13a7d7ff6ee1921c93b6c4