Table writer 6: create directory if not exists (#3414)
Summary:
If the parent directory of the file that a LocalFileSink writes to
does not exist, create the directory first when creating
the LocalFileSink.

Rename FileSink to LocalFileSink, to be more clear.
LocalFileSink is one subclass of the interface
dwio::common::DataSink that implements a local file
system based data sink for a dwio writer.

This is needed by a partitioned table writer, which calls on
the LocalFileSink to write to files under partition subdirectories
that are created during query execution.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3414

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

buck build //fb_presto_cpp:main
buck test velox/dwio/common/tests:

Reviewed By: mbasmanova

Differential Revision: D41723673

Pulled By: gggrace14

fbshipit-source-id: 6aab5e9504f6b4c284f26c85eb664dd415d5db7c