ConstantExpr always allocates a new Vector (#3539)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3539

ConstantExpr will reuse sharedSubexprValues_ and just resize it in evalSpecialForm if it's unique.

Expr will always grab a copy of the shared_ptr through the class's value() function, so this optimization can never be applied
https://github.com/facebookincubator/velox/blob/ad12f0154504b0feb9609aaba767fec670248b10/velox/expression/Expr.cpp#L134

This means ConstantExpr allocates a new Vector every time it's invoked.

To fix this, since evalSpecialForm is allocating a new Vector anyway, I just assign that to sharedSubexprValues_ so that after the first call, assuming nothing downstream is holding onto the pointer, the Vector can be reused.

(I also considered making a copy to return in value() instead, but that meant making a copy everywhere sharedSubexprValues_ could be exposed, and this felt easier to maintain.)

I added a benchmark to demonstrate the difference

The best times over a course of a few runs before:
```
============================================================================
[...]benchmarks/basic/SimpleArithmetic.cpp     relative  time/iter   iters/s
============================================================================
multiplyLiteralSmall                                        5.41ms    185.00
multiplyLiteralMedium                                     665.42us     1.50K
multiplyLiteralLarge                                      286.71us     3.49K
```

The best times over a course of a few runs after:
```
============================================================================
[...]benchmarks/basic/SimpleArithmetic.cpp     relative  time/iter   iters/s
============================================================================
multiplyLiteralSmall                                        4.43ms    225.95
multiplyLiteralMedium                                     522.80us     1.91K
multiplyLiteralLarge                                      269.41us     3.71K
```

The difference is significant for small and medium batches.

Reviewed By: kKPulla

Differential Revision: D42112520

fbshipit-source-id: 3f3b80455120ba1420f42a7d6defdd2eeffc6bf6